move `scale_***` functions around
