+++ src/Configurations.jl
@@ -17,6 +17,9 @@ using StaticArrays, LinearAlgebra, Statistics
 using ..BoundaryConditions
 using ..CustomTypes
 
+import ..BoundaryConditions: scale_xyz, scale_xy, scale_z
+export scale_xyz, scale_xy, scale_z
+
 export Config
 export distance2,
     get_distance2_mat, get_distance2_mat!, get_tan, get_tantheta_mat, get_tantheta_mat!
@@ -230,6 +233,33 @@ function get_tantheta_mat!(dest, config::Config)
         dest[i, i] = 0
     end
     return dest
+end
+
+scale_xyz(vector, α) = α * vector
+function scale_xyz(config::Config, α)
+    return Config(scale_xyz(config.positions, α), scale_xyz(config.boundary_condition, α))
+end
+function scale_xy(pos, scale)
+    new_pos = map(pos) do p
+        SVector(p[1] * scale, p[2] * scale, p[3])
+    end
+    return new_pos
+end
+function scale_xy(config::Config, scale)
+    return Config(
+        scale_xy(config.positions, scale), scale_xy(config.boundary_condition, scale)
+    )
+end
+function scale_z(pos, scale)
+    new_pos = map(pos) do p
+        SVector(p[1], p[2], p[3] * scale)
+    end
+    return new_pos
+end
+function scale_z(config::Config, scale)
+    return Config(
+        scale_z(config.positions, scale), scale_z(config.boundary_condition, scale)
+    )
 end
 
 end
